Squash | UserSettigns logic
* Dev | Закончил работу над дизайном и  запросами Authorization

* Dev | Закончил с дизайном registrationPage

* Update | добавил в валидацию проверку на уникальность login

* Закончил работу с обновлением Authorization

* Dev | Изменение UserSettings

* Dev | Добавил изменения данных user без валидации

* Dev | Закончил работу над UserSettings

+++ AutoCareDiray/ViewModels/MainPage/UserSettingsViewModal.cs
@@ -1,11 +1,13 @@
 ﻿using AutoCareDiray.Models;
 using AutoCareDiray.Service;
 using AutoCareDiray.View;
+using CommunityToolkit.Maui.Alerts;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -14,6 +16,8 @@ namespace AutoCareDiray.ViewModels
     public partial class UserSettingsViewModal : ObservableObject
     {
         private readonly IApiService _apiService;
+        private readonly IDialogService _dialogService;
+        private UserValidation _userValidation;
 
         [ObservableProperty]
         public string nickName;
@@ -21,49 +25,69 @@ namespace AutoCareDiray.ViewModels
         public string email;
         [ObservableProperty]
         public string user_id;
+        [ObservableProperty]
+        public string textError;
 
-        public UserSettingsViewModal(IApiService apiService) 
+        public UserSettingsViewModal(IApiService apiService, IDialogService dialogService) 
         {
             _apiService = apiService;
+            _dialogService = dialogService;
+            _userValidation = new UserValidation(apiService);
             NickName = Preferences.Get("NickName","null");
             Email = Preferences.Get("Email", "null");
-            User_id = Preferences.Get("User_id", "00");
+            User_id = Preferences.Get("User_id", "0");
         }
 
-
+        [RelayCommand]
+        public async void SaveProfil()
+        {
+            if(int.TryParse(User_id,out int result))
+            {
+                var user = new UserDTO
+                {
+                    NickName = NickName,
+                    Email = Email,
+                    User_Id = result
+                };
+                var response = await _apiService.UpdateUserApiAsync(user);
+                TextError = response.Message;
+            }
+        }
         [RelayCommand]
         public async void ExitProfil()
         {
-            Preferences.Remove("User_Id");
-            Preferences.Remove("NickName");
-            Preferences.Remove("Email");
-            Preferences.Remove("is_login");
+            var result = await _dialogService.ShowConfirmationMessage("Вы точно хотите выйти с профиля?");
+            if (result == true)
+            {
+                Preferences.Remove("User_Id");
+                Preferences.Remove("NickName");
+                Preferences.Remove("Email");
+                Preferences.Remove("is_login");
 
-            var authPage = Application.Current.Handler.MauiContext.Services.GetService<AuthorizationPage>();
-            await Shell.Current.Navigation.PushModalAsync(authPage);
-
+                await Shell.Current.GoToAsync("//AuthorizationPage");
+            }
         }
         [RelayCommand]
         public async void DeleteProfil()
         {
-            var UserIdString = Preferences.Get("User_id", null);
-            if (int.TryParse(UserIdString, out var UserId))
+            var result = await _dialogService.ShowConfirmationMessage("Вы точно хотите удалить пользователя?");
+            if (result == true)
             {
-                var response = await _apiService.DeleteUserApiAsync(UserId);
-                if (response.Success == true)
+                var UserIdString = Preferences.Get("User_id", null);
+                if (int.TryParse(UserIdString, out var UserId))
                 {
-                    Preferences.Remove("User_Id");
-                    Preferences.Remove("NickName");
-                    Preferences.Remove("Email");
-                    Preferences.Remove("is_login");
+                    var response = await _apiService.DeleteUserApiAsync(UserId);
+                    if (response.Success == true)
+                    {
+                        Preferences.Remove("User_Id");
+                        Preferences.Remove("NickName");
+                        Preferences.Remove("Email");
+                        Preferences.Remove("is_login");
+
+                        await Shell.Current.GoToAsync("//AuthorizationPage");
+                    }
+                    else TextError = response.Message;
                 }
-
-                var authPage = Application.Current.Handler.MauiContext.Services.GetService<AuthorizationPage>();
-                await Shell.Current.Navigation.PushModalAsync(authPage);
-            }
-            else
-            {
-
             }
         }
     }
